♻️  Refactor: rename queries file to recur.event.repo for clarity
