+++ packages/backend/src/sync/services/notify/handler/gcal.notification.handler.ts
@@ -3,7 +3,7 @@ import { gCalendar } from "@core/types/gcal";
 import { SyncError } from "@backend/common/constants/error.constants";
 import { error } from "@backend/common/errors/handlers/error.handler";
 import gcalService from "@backend/common/services/gcal/gcal.service";
-import { RecurringEventRepository } from "@backend/event/queries/event.recur.queries";
+import { RecurringEventRepository } from "@backend/event/services/recur/repo/recur.event.repo";
 import { getSync, updateSyncTokenFor } from "@backend/sync/util/sync.queries";
 import { Summary_Sync } from "../../../sync.types";
 import { GcalSyncProcessor } from "../../sync/gcal.sync.processor";
