+++ packages/backend/src/sync/services/sync/gcal.sync.processor.test.ts
@@ -18,7 +18,7 @@ import {
 import { simulateDbAfterGcalImport } from "@backend/__tests__/helpers/mock.events.init";
 import { createRecurrenceSeries } from "@backend/__tests__/mocks.ccal/ccal.mock.db.util";
 import { mockGcalEvent } from "@backend/__tests__/mocks.gcal/factories/gcal.event.factory";
-import { RecurringEventRepository } from "@backend/event/queries/event.recur.queries";
+import { RecurringEventRepository } from "@backend/event/services/recur/repo/recur.event.repo";
 import { isInstance } from "@backend/event/services/recur/util/recur.util";
 import { Change_Gcal } from "@backend/sync/sync.types";
 import { GcalSyncProcessor } from "./gcal.sync.processor";
